+++ lib/src/flutter/widget/widget.dart
@@ -1,2 +1,5 @@
 export 'button.dart';
+export 'checkbox.dart';
+export 'radio.dart';
+export 'switch.dart';
 export 'text_form_field.dart';